Add `RELOCATION_ID` macro for upstream compatibility.
This macro was removed early in NG development since it needed a
dynamic system for switching between SE and AE IDs, but it is possible
to still have the macro as an alias to the current `REL::RelocationID`
class. This helps maintain compatibility with the powerof3 fork, so
codebases can switch between it and this one more seamlessly.