Add raw form ID handling in `TESForm` and `TESDataHandler` distinct from local form IDs.
These functions are intended to safely handle raw form ID use even when dealing with
ill-formed mods. The existing `GetLocalFormID()` function ignores the index bits
entirely, assuming that the form ID refers to one defined within a target `TESFile`.
However, this is not the case when a mod "overrides" a form that does not exist in the
master. In that case, SSE will create the form and makes it owner file (the result of
`GetFile(0)`) the mod with the bad override, but the global form ID will still use the
compile index of the master that was expected to have the original form. To lookup the
form by its raw ID still requires going to the mod with the bad override, but it is
necessary to distinguish it's index as being that of a master and not the mod itself.