Fix Windows header include fragility, enable fmt/spdlog header-only for downstreams.
A common issue with new development on CommonLibSSE is that any Windows headers
must be included after the full set of CLibSSE headers. This can be an issue
if the end user is not aware of the limitation, or due to a common problem
with the Vcpkg port which causes it to erroneously use the header-only
deployment of spdlog, which is included through CommonLibSSE headers and in
header-only mode includes Windows headers.