Add `no_init` param to inventory getters
The implementation for GetInventoryChanges in CLIB does not match that of the game; the game's does not attempt to initialize the inventory if it hasn't been already.

Initializing the inventory can have side-effects, like initializing the leveled items. This is undesirable in many circumstances, so we add a `no_init` parameter to any function that gets the inventory to prevent this.

I've set `no_init` to false by default to prevent breaking any mod relying on the current behavior